[IMP] web: convert mobile control_panel test to hoot
task-4028335
