+++ addons/web/static/tests/search/control_panel.test.js
@@ -1,4 +1,4 @@
-import { expect, test } from "@odoo/hoot";
+import { expect, test, getFixture } from "@odoo/hoot";
 import { click, press, queryAll } from "@odoo/hoot-dom";
 import { animationFrame } from "@odoo/hoot-mock";
 import { reactive } from "@odoo/owl";
@@ -160,4 +160,34 @@ test("view switcher hotkey cycles through views", async () => {
     press(["alt", "shift", "v"]);
     await animationFrame();
     expect(`.o_list_view`).toHaveCount(1);
+});
+
+test.tags("mobile")("Control panel is shown/hide on top when scrolling", async () => {
+    await mountWithSearch(
+        ControlPanel,
+        { resModel: "foo" },
+        {
+            viewSwitcherEntries: [
+                { type: "list", active: true, icon: "oi-view-list", name: "List" },
+                { type: "kanban", icon: "oi-view-kanban", name: "Kanban" },
+            ],
+        }
+    );
+    const contentHeight = 200;
+    const sampleContent = document.createElement("div");
+    sampleContent.style.minHeight = `${2 * contentHeight}px`;
+    const target = getFixture();
+    target.appendChild(sampleContent);
+    target.style.maxHeight = `${contentHeight}px`;
+    target.style.overflow = "auto";
+    target.scrollTo({ top: 50 });
+    await animationFrame();
+    expect(".o_control_panel").toHaveClass("o_mobile_sticky", {
+        message: "control panel becomes sticky when the target is not on top",
+    });
+    target.scrollTo({ top: -50 });
+    await animationFrame();
+    expect(".o_control_panel").not.toHaveClass("o_mobile_sticky", {
+        message: "control panel is not sticky anymore",
+    });
 });
